third_party/WebKit/LayoutTests/plugins/return-negative-one-from-write-expected.txt

This tests that NPP_DestroyStream is called with NPRES_NETWORK_ERR not called if a plugin returns -1 its NPP_Write callback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS plugin.returnNegativeOneFromWrite is true
PASS successfullyParsed is true

TEST COMPLETE
PASS newStreamCalled is false
PASS newStreamCalled is true
PASS reason is 1

TEST COMPLETE
